fix(regex): update input placeholders in isEmailDemo.vue and isIdCardDemo.vue
The input placeholder in isEmailDemo.vue was changed from "Enter phonenum..." to "Enter email...". This change provides a more accurate description of the expected input.

The input placeholder in isIdCardDemo.vue was changed from "Enter phonenum..." to "Enter Id card...". This change provides a more accurate description of the expected input.