feat(docs): add IsArrayEqualDemo component to the Vitepress theme
A new file `isArrayEqualDeml.vue` was added to the `docs/.vitepress/theme/components/array` directory. This component imports the `isArrayEqual` function from the `utils-snap-fn` package and demonstrates its usage by comparing two arrays `[1, 2, 3]` and `[1, 2, 3]`. The result of the comparison is displayed in the template.

In addition, the `IsArrayEqualDemo` component was registered in the `index.ts` file of the Vitepress theme, allowing it to be used in the documentation.