feat(docs): add RandomUUIDDemo component to the Vitepress theme
A new file `randomUUIDDemo.vue` has been added to the `docs/.vitepress/theme/components/random` directory. This component generates a random UUID using the `generateUUID` function from the `utils-snap-fn` package and displays it on the page.

In addition, the `RandomUUIRDemo` component has been imported and registered in the `index.ts` file of the Vitepress theme, allowing it to be used in the documentation.